add plugin contributor's guide
