+++ community/plugins/spring-ai-alibaba-starter-plugin-baidusearch/src/main/java/com/alibaba/cloud/ai/plugin/baidu/BaiduSearchPluginConfiguration.java
@@ -1,0 +1,45 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.alibaba.cloud.ai.plugin.baidu;
+
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Description;
+
+/**
+ * .
+ *
+ * @author: KrakenZJC
+ * @since : 2024-11-18
+ **/
+
+@Configuration
+@ConditionalOnClass(BaiduSearchService.class)
+public class BaiduSearchPluginConfiguration {
+    
+    @Bean
+    @ConditionalOnMissingBean
+    @Description("Use baidu search engine to query for the latest news.") // function
+    // description
+    public BaiduSearchService baiduSearchService() {
+        return new BaiduSearchService();
+    }
+    
+}
