+++ community/plugins/spring-ai-alibaba-starter-plugin-time/src/main/java/com/alibaba/cloud/ai/plugin/time/TimeAutoConfiguration.java
@@ -1,0 +1,27 @@
+package com.alibaba.cloud.ai.plugin.time;
+
+import com.alibaba.cloud.ai.plugin.time.service.GetCurrentLocalTimeService;
+import com.alibaba.cloud.ai.plugin.time.service.GetCurrentTimeByTimeZoneIdService;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Description;
+
+@AutoConfiguration
+public class TimeAutoConfiguration {
+
+	@Bean
+	@ConditionalOnMissingBean
+	@Description("Get the current local time")
+	public GetCurrentLocalTimeService getCurrentLocalTimeFunction() {
+		return new GetCurrentLocalTimeService();
+	}
+
+	@Bean
+	@ConditionalOnMissingBean
+	@Description("Get the time of a specified city")
+	public GetCurrentTimeByTimeZoneIdService getCityTimeFunction() {
+		return new GetCurrentTimeByTimeZoneIdService();
+	}
+
+}
