+++ community/plugins/spring-ai-alibaba-starter-plugin-time/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+com.alibaba.cloud.ai.plugin.time.TimeAutoConfiguration
